Letter to my Future Self
I've been thinking recently about the idea of self-addressing letters,
intending for them to be opened years down the line. They're ideas
frozen in time, unchanged since the time of writing. They represent not
only the environment in which they were written but also the
expectations of what's to come. There's an element of uncertainty,
since we don't know how things are going to turn out, but that
uncertainty isn't a bad thing, necessarily; instead, it invites us to
use our imaginations and dream about the future.

Right now, I'm imagining what this project is going to look like down
the line, and you're reading this commit message.

If you're reading this on GitHub, that last sentence probably didn't
surprise you. But if this project turns out how I'm envisioning it, you
won't be reading this on GitHub, you'll be reading this on my website.

The idea is to build an interface for a project's commit history that
treats them like more than just a changelog. The commit history is a
journal, an immutable record of a project's progress. It's a letter to
the future self (or selves) that will be reading this.

As such, this project will let commit messages take center stage, as if
they're posts on a blog. The file changes will be there too
(hopefully), but shouldn't take away from the story being told by the
commit messages. And what better project to test this idea out on than
itself?

Some project goals:
- Commit messages are displayed in order from oldest to newest, with
 the author and date displayed alongside each entry. If you're reading
 this, success!
- Some kind of rich text support for messages. Maybe not full Markdown,
 but hopefully *this is bolded*, _this is underlined_, and maybe
 [this](https://lemoing.ca/) is a link? This list may even be bulleted.
 Who knows.
- Some kind of "files changed" indicator, even if it's not a full
 GitHub-style diff comparison. Should be accessible without interrupting
 the flow of the commit messages.

I have no idea what else it'll look like. I won't be able to edit any
of this after the fact if this doesn't turn out like I described, but
that's okay. That's the uncertainty I was talking about earlier.

I'm writing this initial commit message, before anything is built or
shared, with the hope this will eventually be the first chapter of this
project's story. And that you'll be reading it, using this project.

It's a Saturday afternoon in September 2020. I'm in Kitchener, Ontario,
and the weather is brisk but sunny. Remember to drink water and take a
break from your computer more often than you think you need to.

Signing off the initial commit.