Adding some DOM boilerplate
When I'm working in a web frontend role, I usually use React. But
there's no React on my website, and it's not usually the tool I reach for
when I'm working on side projects. I think React is an incredible tool
and I do like working with it, but there is something so nice about not
having to worry about your build configuration. When you write a
JavaScript file and include it in your HTML, it just works. If all that
means is having to add some extra helpers that React would take care of for
you, then for me it's worth the tradeoff.

Here, I'm adding some functions that will help me add DOM elements. This
page won't be fancy, but it will load content dynamically from the
GitHub API, so these will come in handy later. Hopefully that'll mean
allowing us to display other projects besides this one, but we'll see!