Accessing the GitHub API
One mildly interesting thing I've noticed about technology is that
things are almost never ordered from oldest to newest. You're almost
always being shown the new stuff first. It's most noticeable on social
media and blogs, but it's also true on GitHub,which lists commits from
newest to oldest. It's usually the most useful option, but for our
purposes it cuts out the story. You wouldn't start reading a book from
the last page.

Using the git CLI, you can run `git log --reverse` to get a list of
commits from oldest to newest. Unfortunately, at the time of writing,
the GitHub API doesn't offer that option. However, we can find the
number of commits on a branch by using the `Link` header. The header has
pagination information which includes a link to fetch the last page. By
setting the page size to 1, the page parameter in the "last page" link
will be equal to the number of commits on the branch.

It's deceptively difficult to reverse the order of REST API if it
doesn't have a sorting option. Let's say we want to display 10 commits
at a time. When I wrote this, my resume/website repo had 114 commits,
which means grabbing the last page of commits will return 4 items with a
page size of 10. However, increasing the page size won't necessarily
increase the number of commits on the last page, since the number of
commits on the last page is simply the amount left over when dividing
the number of commits by the page size. The remainder of 114 when
divided by 10 is 4, but the remainder of 114 when divided by 11 is also
4!

We could keep increasing the page size in our request until the last
page is big enough for the amount we want to display, but I found the
strategy I ended up using was easier to reason about. We'll fetch both
the last and second last pages; combined, these give us the commits we
need.