Reckoning with the Past
There comes a time in every project where you have to take a look at the
direction you set out when starting off and evaluate those initial goals
with wiser, wearier eyes. Reducing the scope is important because it
keeps the goals tangible, within reach. You almost always make some
concessions or changes in direction. Unfortunately for this project, my
initial goals will be the first thing you see. They're the first thing I
see, now that we can fetch the commits from GitHub.

I can't believe I said I was going to support rich text commit messages.

Linux has it right. Everything I see from them, from their mailing lists
to their README, is glorious, unformatted plain text. Focus on the
message, not the presentation.

Unfortunately for me, this is a quasi-creative-project as well as a
project-project, so presentation matters. I'm hoping the fact that these
are commit messages is a surprise when you first load the page. In order
to best deliver that surprise, there should be some element of disguise.
I need to make this look like something it's not, like a blog post. And
it's hard to make this look like a *convincing* blog post without doing
some kind of text formatting, or at the very least, putting paragraphs
in <p> tags. Which is what I'll start with.

I'm going to be loosely following the
[CommonMark](https://commonmark.org/) spec, but won't implement it all.
They do have a reference implementation that I could use, but I think it
might cool for you all to see that even a bad programmer can write a
parser. Also, I made a mistake in my original message when saying what
the rich text should look like, so maybe I'll have to go rogue and
create yet another Markdown flavour.

One of the original thoughts I had for this project was that it could be
used for coding walkthroughs or tutorials. A lot of the time, these are
written as blog posts anyways. With this project, you'll be able to see
the narrative via commit messages alongisde the actual code diffs, and
if all goes well, it'll even be formatted nicely. So this can double as
a trial for that thought and serve as a walkthrough for how to do all
these things from the ground up.

With that in mind, here's what I'm doing to start off. First, I'm going
through the messages line by line and building up an Abstract Syntax
Tree (AST) containing the structure of the document. Blank lines work to
separate the text into paragraphs; the comments have more info about how
these are handled. Later, we'll add in support for handling lists,
emphasis, and links. For now, our AST looks something like this:

- document
  - paragraph
    - line
    - line
  - paragraph
    - line

Once we have the AST, we can convert it to HTML by
going through it recursively and handling each node
type. For lines, we'll create text nodes; for
paragraphs, we'll create paragraph nodes, and for
the document, we'll use a document fragment with
the paragraphs as children. And that's it, for now!
Next up is tackling everything else.