The False Dichotomy of Recursion vs. Iteration
I've got some explaining to do.

You may be looking at my message parser code and scratching your head. I
know I was when I came back to it after a few days away from the
project. It combines recursion and iteration in a way that can melt your
brain a little bit if you don't know why the decision was made in the
first place. To help you understand why, I'll take you back to the very
beginning.

My first exposure to programming was CS135, the introductory programming
course at the University of Waterloo. It's based around the [teaching
languages included in
DrRacket](https://docs.racket-lang.org/drracket/htdp-langs.html), which
are stripped-down versions of the Racket programming language; so
stripped down, in fact, that they don't support any kind of mutable
state or side effects.

One side effect of this is that I was much more comfortable with
recursion. I didn't learn what a `for` loop until my second semester of
university. It was only after I spent a lot more time programming that I
got comfortable with iteration as well. At that point, both recursion
and iteration were tools in my toolbox, and I could choose to use one,
or both, depending on the problem I was faced with.

So, why mix the two here? Are you even allowed to mix the two? Would it
be easier to pick one or the other?

Not necessarily. Or at least, not for me, for this particular problem.
There's sometimes the sentiment that you have to pick one or the other,
but both lend themselves well to different scenarios, and in this case,
both scenarios were present in the parsing problem.

I came up with this approach when I realized that the system should have
some idea of the context it's in. Depending on the context, a list item
could be a sub-list of an existing list, a sibling list item, or the
start of an entirely new list.

One way to store that context is a stack. Each time you enter a new
context, you can push some data onto the stack and parse, knowing that
the old context will be available afterwards. But recursive functions
have a built-in stack, and you don't have to worry as much about
managing state, or whether you need to break out of your loop or
*shudders* use a `goto` statement to get out of the control flow mess
you made. All you need to do is call a new function and you'll be
magically warped to a new context knowing that your old context is well
taken care of. And when you're done, just return! There's no need to
check if your stack is empty or to worry about what would happen if you
don't.

If recursion is so great, why bother having iteration? The reason
iteration is used here is because within a context, we often need to
check multiple lines. For example, this paragraph spans multiple lines,
so our program will parse all of them within the same context. You can
do this with recursion as well, but for me personally, it was much
easier to run through lines that belong to the same context using
iteration rather than try to build the list recursively. This way, the
message parsing code is a reflection of the data structure we're
building with it, since each function call represents the creation of a
new node in the AST, and take care of everything that isn't a new node
(new lines in the same paragraph, for example) with iteration.

So, with the recursive/iterative framework, we're able to add list
parsing without too much of a hitch. Lists are naturally recursive (you
can have a list within a list within a...) and list items can also span
multiple lines, so it's fairly straightforward to add them for the
reasons described above.

This did reveal a bug in the earlier code where I was skipping some
lines when exiting from a context. This was fixed by returning one less
than the current index, which tells the parent context that the line
should be parsed again. The regex test for deciding if a line is a list
item is also pretty hairy; see the comments for more details.

Finally, there are some differences between my list implementation and
Markdown, mostly because there are so many edge cases that
industry-standard parsers need to consider that I didn't want to have to
deal with at this point. I also plan on adding the option to toggle rich
text parsing on and off at some point in the future which will help if
any incompatibilities make a mess of someone's formatting.

Next up is parsing the in-line content! This includes:
1. Emphasis (which I made a mistake specifying in the *very first
commit* - oops!)
2. `Code fences`
3. [Links](https://lemoing.ca/)