And just like that, we have rich text!
After the embarassing amount of time I spent trying to figure out lists,
wrapping up rich text was surprisingly easy. The features we added this
time around are all inline elements, meaning they apply to individual
snippets of text rather than whole sections. As a result, we only need
to touch the `'line'` AST nodes. And overall, it's not a whole lot
different from the work we did earlier; the main difference is that
we're going character by character instead of line by line.

The other thing to note is it isn't very robust. If you read the
[CommonMark Spec on inline content](https://spec.commonmark.org/0.29/#inlines),
they go into a ton of around how edge cases are treated. This is due in
part to the ambiguity of Markdown; the same text can produce different
results based on the parser you use. However, in the interest of getting
something quick ready for this project, where Markdown is just a way to
make the end result seem more like a blog post, this approach is fine.
I'll just try not to accidentally put emphasis in my code blocks.

At a later date, we may try to refactor the message parsing a bit to
make it more readable. It's not the most straightforward code I've ever
written, but hopefully these hyper-detailed commit messages make up for
that. I'll leave that as a TODO for now and move on to the final element
of blog-post-like disguise: styling and layout.