A Big Milestone
It's hard to call a project done, especially when there's still a few
meaningful additions you can think of making. But this project has
arrived at a state where I could leave it for some time and still be
content with what I've built.

I wouldn't consider styling one of my strong suits, but even the small
amount I added was enough to evoke a somewhat-convincing blog-post-like
feeling. That, combined with a few changes to navigation, layout, and
presentation, have left it feeling crisp, if not complete. Obviously
I'll have to do some more testing when I try to launch, so watch for a
commit shortly after this one where I fix all the mistakes I'm sure I've
made but haven't spotted yet.

This may be good enough for a beta release, but there's a few more
things I'd like to add before I use the 1.0 label.
- My original goal of having an indicator of files changed is still on
my radar, but I'm worried it'll be even more of a rabbit hole as rich
text was. For now, I'm sheepishly deferring to GitHub to display that
info for me.
- It would be good to see if I can be a bit smarter about storing some
of the stuff we've done in storage. This would let me toggle rich text
on and off, cache API responses, and save the scroll position; right
now, if you navigate to the commit's changes page, your scroll position
will reset when you come back since all the DOM content gets
regenerated. This wouldn't happen if I had made a static site, but this
way you can put your own projects in to see what they look like.
- Speaking of trying out your own projects, I also want to add a way to
easily load other repos in. For now, the search params `owner` and
`repo` allow you to load other projects in besides this one: add the
search string `?owner={owner}&repo={repo}` to the end of the URL to load
a public repo that lives at `https://github.com/{owner}/{repo}`.

Even with all of those potential future improvements, I'm proud of where
this project is at this point in time. When I first had the idea for the
project, I was obsessed with the idea of a project journal, a kind of
living document that evolves and changes as the project progresses. And
as I read through the entries of the project again and again, I'm
constantly reminded of the progress I've made, as well as the progress I
still want to make. It's motivating, in a way, seeing it spelled out so
clearly.

It's now October. The weather is a little colder and there are a few
more leaves on the ground. But not much time has passed since I started
this off. I still need reminders to drink more water and take breaks
from my computer. But overall, things are going well, and I'm excited to
see where things go.

See you soon, and take care.