Stepping Stones
When I'm working on a project, I tend to dream pretty big. It's easy to
get so caught up in thinking about all the cool things that *could* be
added that all the intermediate steps lose their shine a little.

This commit represents one of those intermediate steps. It's not flashy
or exciting. It's the glue that connects another part of the GitHub API
so that we know what files changed, and how they changed. It's a
stepping stone for what is going to come later, which I hope will be
really cool; I'll get to talk about how git patches work, and how we can
translate them into a diff that'll let you follow along with the code
I'm writing. I might even try to write an adjacent project to show how
this could work for a code walkthrough. But to get there, I have to go
through this step first.

There's other things on my list of `TODO`s that are even less exciting;
stuff that won't result in any visible change to the project but *may*
improve the structure, or make it run slightly more smoothly. Those
things tend to get punted for as long as they aren't critical.

But it is very important to spend the time on these less glamorous
steps. I've had lofty goals for projects before, only to get discouraged
and distracted when the intermediate stages don't look as polished or
complete as I had envisioned in my head. These steps are just as
important, if not more so, since they tie the project together.

So, although a button with some info about files changed isn't
necessarily exciting in and of itself, it's an important stepping stone
towards the end goal, and that's exciting too.