Patching things up
Have you ever wondered GitHub, VSCode, or commands like `git diff`
display changes to source code? I had no idea before starting this
project, but it's actually not very complicated.

As part of the API response for individual commits, GitHub returns a
list of files changed. Most files[1] have an associated `patch`
property, which is a long string that represents all the changes that
were applied to that file. This string contains all the info we need to
display a diff.

The patch string is broken up into chunks, separated by header lines.
The header tells us the start line and the number of lines changed for
both the LHS and RHS. In this case, the LHS is the previous commit and
the RHS is the new commit. A header line looks like `@@ -{LHS Start
Line},{LHS Lines Changed} +{RHS Start Line},{RHS Lines Changed}`.

After each header comes the actual line changes. Each of these is on a
new line that can start with either a space, a minus (`'-'`) or a plus
(`'+'`). Lines starting with a space are unchanged in both files but are
included as context. Intuitively, lines starting with a minus are
removed and lines starting with a plus are added.

Sometimes, the number of lines removed doesn't equal the number of lines
added. Because of this, we need to figure out how lines in the LHS map
to lines in the RHS. The way to determine this is straightforward: if a
removal line is followed by an addition line, those two lines are paired
up. Any other removed or added lines that don't match this pattern are
treated as standalone lines.

With all this in mind, we can parse the whole patch by first looking for
header lines, and then parsing the associated changes for each header.
There's some more detail in the comments, but basically, this parser
will output an array of changes for each file and write it to the
console.

The next step is turning these changes into HTML.

[1] I believe some files are missing patches because they don't have a
file type that GitHub recognizes as being text. This happens for binary
files, like images, but I've also seen it happen on some text files. Not
exactly sure what causes this; could be that the extension isn't
recognized or the patch is too big to send. In any case, for most files,
it's not a problem.