Another Milestone!
It sure is weird seing the first code I wrote for this project displayed
alongside these messages. The code for displaying the patches is finally here!

At first, I thought I had produced some kind of mutant quine[1] (a
Franken-quine, if you will) that outputs not only its own source, but
also the list of all of the changes ever made it its own source. Crazy.
Unfortunately, this doesn't count as a quine since we're receiving input
from GitHub. Still crazy, in my opinion.

I've actually had this code written out for a while. I'm only now
getting around to publishing it because it's one of those things that's
so easy to keep tweaking and fiddling with. This was the last element on
my original checklist and I was constantly trying to reconcile my
original vision with what I was actually able to produce.

There's another reason I haven't released this yet, if I'm honest;
writing the commit messages is hard work too. Sometimes I have a good
idea of what to write, but that wasn't the case this time. Just like
with the code, I've written and rewritten a lot more for these messages
than you'll see in the finished project.

Sometimes, what does get me to finally write things out is the desire to
implement another feature, which I can't do until I wrap this section up
as well. So, counter-intuitively, having to write both things in
parallel can lead to *less* writer's block, not more.

Anyways, I'm happy to cap off another milestone. There are still several
improvements to make, but on the whole they're mostly minor compared to
the features that are already in place. For the first time since
starting this project, I don't have anything else big in mind. There
are, however, many small things; for starters, this is still basically
unreadable on mobile, so look for a fix to that next!

It's now November, and it might be the end of the last weekend of fall
weather. It snowed for the first time today, and while it won't stick,
it definitely marks the beginning of the end. I also did yoga for the
first time in months. If I'm going to survive this winter inside, I'll
need to make that more of a habit; we'll see if that actually happens.

Until next time!

[1] A quine is a program that takes no input and produces an exact copy
of its own source code. See
[here](http://www.madore.org/~david/computers/quine.html) for more.